add progress to all states
