+++ app/src/main/java/org/stepik/android/presentation/step_quiz_review/reducer/StepQuizReviewReducer.kt
@@ -40,10 +40,10 @@ constructor() : StateReducer<State, Message, Action> {
                     is State.Loading -> {
                         val quizState = createAttemptLoadedState(message.sessionData.attempt, message.sessionData.submission)
                         val newState =
-                            if (message.progress != null) {
-                                State.Completed(quizState, message.progress, message.instruction)
+                            if (message.sessionData.session.isFinished) {
+                                State.Completed(quizState, message.sessionData.session, message.instruction, message.progress)
                             } else {
-                                State.SubmissionSelected(quizState, message.sessionData.session, message.instruction)
+                                State.SubmissionSelected(quizState, message.sessionData.session, message.instruction,  message.progress)
                             }
 
                         newState to emptySet()
@@ -57,6 +57,7 @@ constructor() : StateReducer<State, Message, Action> {
                     is State.Loading ->
                         State.SubmissionNotMade(
                             quizState = message.quizState,
+                            progress = message.progress,
                             instruction = message.instruction
                         ) to emptySet()
                     else -> null
@@ -72,7 +73,7 @@ constructor() : StateReducer<State, Message, Action> {
             is Message.SolveAgain ->
                 when (state) {
                     is State.SubmissionNotSelected ->
-                        State.SubmissionNotMade(state.quizState, state.instruction) to emptySet()
+                        State.SubmissionNotMade(state.quizState, state.instruction, state.progress) to emptySet()
 
                     else -> null
                 }
@@ -83,7 +84,7 @@ constructor() : StateReducer<State, Message, Action> {
                         val stepQuizViewState =
                             state.quizState.copy(attempt = message.attempt, submissionState = StepQuizView.SubmissionState.Loaded(message.submission))
 
-                        State.SubmissionNotSelected(stepQuizViewState, state.instruction) to emptySet()
+                        State.SubmissionNotSelected(stepQuizViewState, state.instruction, state.progress) to emptySet()
                     }
 
                     else -> null
@@ -97,7 +98,7 @@ constructor() : StateReducer<State, Message, Action> {
                             ?.id
 
                         if (submissionId != null) {
-                            State.SubmissionSelectedLoading(state.quizState, state.instruction) to setOf(Action.CreateSessionWithSubmission(submissionId))
+                            State.SubmissionSelectedLoading(state.quizState, state.instruction, state.progress) to setOf(Action.CreateSessionWithSubmission(submissionId))
                         } else {
                             null
                         }
@@ -109,7 +110,7 @@ constructor() : StateReducer<State, Message, Action> {
             is Message.CreateSessionError ->
                 when (state) {
                     is State.SubmissionSelectedLoading ->
-                        State.SubmissionNotSelected(state.quizState, state.instruction) to setOf(Action.ViewAction.ShowNetworkError)
+                        State.SubmissionNotSelected(state.quizState, state.instruction, state.progress) to setOf(Action.ViewAction.ShowNetworkError)
 
                     else -> null
                 }
@@ -117,7 +118,7 @@ constructor() : StateReducer<State, Message, Action> {
             is Message.SessionCreated ->
                 when (state) {
                     is State.SubmissionSelectedLoading ->
-                        State.SubmissionSelected(state.quizState, message.reviewSession, state.instruction) to emptySet()
+                        State.SubmissionSelected(state.quizState, message.reviewSession, state.instruction, state.progress) to emptySet()
 
                     else -> null
                 }
