--- app/src/main/java/org/stepik/android/view/step_quiz_review/ui/delegate/StepQuizReviewInstructorDelegate.kt
@@ -1,13 +1,0 @@
-package org.stepik.android.view.step_quiz_review.ui.delegate
-
-import android.view.View
-import kotlinx.android.extensions.LayoutContainer
-import org.stepik.android.presentation.step_quiz_review.StepQuizReviewView
-
-class StepQuizReviewInstructorDelegate(
-    override val containerView: View
-) : LayoutContainer, StepQuizReviewDelegate {
-    override fun render(state: StepQuizReviewView.State) {
-
-    }
-}
