+++ app/src/main/res/values-ru/plurals.xml
@@ -137,7 +137,7 @@
     </plurals>
 
     <plurals name="step_quiz_review_given_completed">
-        <item quantity="one">%d рецензию отправлена.</item>
+        <item quantity="one">%d рецензия отправлена.</item>
         <item quantity="few">%d рецензии отправлены.</item>
         <item quantity="many">%d рецензий отправлены.</item>
         <item quantity="other">%d рецензий отправлены.</item>
@@ -158,7 +158,7 @@
     </plurals>
 
     <plurals name="step_quiz_review_taken_completed">
-        <item quantity="one">%d рецензию получена.</item>
+        <item quantity="one">%d рецензия получена.</item>
         <item quantity="few">%d рецензии получены.</item>
         <item quantity="many">%d рецензий получены.</item>
         <item quantity="other">%d рецензий получены.</item>
